requirements-test.txt
#-r requirements.txt
django-nose==1.4.6
factory-boy==2.5.2
fake-factory==0.5.2

flake8==2.3.0
coverage==3.7.1
requests-mock==1.5.2
beautifulsoup4==4.4.1

django-environ==0.3.0

psycopg2==2.7.6

mock>=1.0.1

